fix location security

+++ app/Services/LocationService.php
@@ -4,14 +4,14 @@ namespace App\Services;
 
 use App\Models\Location;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 
 class LocationService
 {
   public function store($data): RedirectResponse
   {
-    $location = new Location();
-    $location::create([
+    Location::create([
       'user_id' => auth()->id(),
       'name' => $data->get('name'),
       'longitude' => $data->get('longitude'),
@@ -21,23 +21,30 @@ class LocationService
     return redirect()->route('dashboard.get')->with('success', '👍 Added.');
   }
 
-  public function edit($id): View
+  public function edit($id): View|RedirectResponse
   {
     $location = Location::findOrFail($id);
-    return view('location-edit')->with(['location' => $location]);
+    if (Gate::denies('update', $location)) {
+      return redirect()->route('dashboard.get')->with('error', '😠 You do not have access.');
+    }
+    return view('location-edit', ['location' => $location]);
   }
 
   public function update($id, $data): RedirectResponse
   {
-
     $location = Location::findOrFail($id);
+    if (Gate::denies('update', $location)) {
+      return redirect()->route('dashboard.get')->with('error', '😠 You do not have access.');
+    }
     $location->update($data->validated());
-
     return redirect()->route('dashboard.get')->with('success', '👍 Updated.');
   }
 
   public function destroy(Location $location): RedirectResponse
   {
+    if (Gate::denies('delete', $location)) {
+      return redirect()->route('dashboard.get')->with('error', '😠 You do not have access.');
+    }
     $location->delete();
     return redirect()->route('dashboard.get')->with('success', '👍 Deleted.');
   }
